Copy AOFlagger flags to dedicated column and reuse this column unless rerun_aoflagger option is passed.

+++ tabascal/utils/flag_data.py
@@ -26,15 +26,9 @@ def write_perfect_flags(ms_path: str, n_sigma: float = 3.0):
         flags = xr.zeros_like(xds.DATA).astype(bool)
 
     xds = xds.assign(FLAG=flags)
-    dask.compute(
-        xds_to_table(
-            [
-                xds,
-            ],
-            ms_path,
-            ["FLAG"],
-        )
-    )
+    cols = ["FLAG"]
+
+    dask.compute(xds_to_table([xds], ms_path, cols))
 
     print()
     print(f"Flag Threshold : {n_sigma: .1f} sigma")
@@ -43,49 +37,87 @@ def write_perfect_flags(ms_path: str, n_sigma: float = 3.0):
     return flags
 
 
+def write_to_aoflags(ms_path: str):
+
+    xds_ms = xds_from_ms(ms_path)[0]
+
+    xds_ms = xds_ms.assign({"AO_FLAGS": xds_ms["FLAG"]})
+    cols = ["AO_FLAGS"]
+
+    print("Writing AOFlagger flags to 'AO_FLAGS' column in MS file.")
+
+    dask.compute(xds_to_table([xds_ms], ms_path, cols))
+
+
+def write_aoflags_to_flag(ms_path: str):
+
+    xds_ms = xds_from_ms(ms_path)[0]
+
+    xds_ms = xds_ms.assign(FLAG=xds_ms["AO_FLAGS"])
+    cols = ["FLAG"]
+
+    print("Writing AOFlagger flags to 'FLAG' column in MS file.")
+
+    dask.compute(xds_to_table([xds_ms], ms_path, cols))
+
+
 def run_aoflagger(
     ms_path: str,
     data_column: str = "DATA",
     strategy_paths: list = None,
     sif_path: str = None,
     bash_exec: str = "/bin/bash",
+    rerun_aoflagger: bool = False,
 ):
 
     if ms_path[-1] == "/":
         ms_path = ms_path[:-1]
 
-    data_dir, ms_file = os.path.split(os.path.abspath(ms_path))
+    if not rerun_aoflagger:
+        xds_ms = xds_from_ms(ms_path)
+        try:
+            aoflags = xds_ms["AO_FLAGS"]
+            write_aoflags_to_flag(ms_path)
+            print("Using previous AOFlagger run.")
+        except:
+            aoflags = None
 
-    docker_opts = "--rm -v /etc/group:/etc/group -v /etc/passwd:/etc/passwd -v /etc/shadow:/etc/shadow -v/etc/sudoers.d:/etc/sudoers.d -e HOME=${HOME} --user=`id -ur`"
-    container_cmd = f"docker run {docker_opts} -v {data_dir}:/data --workdir /data stimela/aoflagger:latest"
+    if not aoflags and rerun_aoflagger:
 
-    if sif_path is not None:
-        sif_path = os.path.abspath(sif_path)
-        container_cmd = (
-            f"singularity exec --bind {data_dir}:/data --pwd /data {sif_path}"
-        )
+        data_dir, ms_file = os.path.split(os.path.abspath(ms_path))
 
-    if strategy_paths is not None:
-        write_perfect_flags(ms_path, 0)
-        for strategy_path in strategy_paths:
+        docker_opts = "--rm -v /etc/group:/etc/group -v /etc/passwd:/etc/passwd -v /etc/shadow:/etc/shadow -v/etc/sudoers.d:/etc/sudoers.d -e HOME=${HOME} --user=`id -ur`"
+        container_cmd = f"docker run {docker_opts} -v {data_dir}:/data --workdir /data stimela/aoflagger:latest"
 
-            strategy_path = os.path.abspath(strategy_path)
-            shutil.copy(strategy_path, data_dir)
-            strategy_file = os.path.split(strategy_path)[1]
-            strategy = f"-strategy /data/{strategy_file}"
+        if sif_path is not None:
+            sif_path = os.path.abspath(sif_path)
+            container_cmd = (
+                f"singularity exec --bind {data_dir}:/data --pwd /data {sif_path}"
+            )
 
-            aoflag_cmd = f"{container_cmd} aoflagger -column {data_column} {strategy} /data/{ms_file}"
+        if strategy_paths is not None:
+            write_perfect_flags(ms_path, 0)
+            for strategy_path in strategy_paths:
+
+                strategy_path = os.path.abspath(strategy_path)
+                shutil.copy(strategy_path, data_dir)
+                strategy_file = os.path.split(strategy_path)[1]
+                strategy = f"-strategy /data/{strategy_file}"
+
+                aoflag_cmd = f"{container_cmd} aoflagger -column {data_column} {strategy} /data/{ms_file}"
+                subprocess.run(aoflag_cmd, shell=True, executable=bash_exec)
+
+            print()
+            print(
+                f"Strategies : {[os.path.split(strategy)[1] for strategy in strategy_paths]}"
+            )
+
+        else:
+            aoflag_cmd = f"{container_cmd} aoflagger /data/{ms_file}"
             subprocess.run(aoflag_cmd, shell=True, executable=bash_exec)
+            print()
 
-        print()
-        print(
-            f"Strategies : {[os.path.split(strategy)[1] for strategy in strategy_paths]}"
-        )
-
-    else:
-        aoflag_cmd = f"{container_cmd} aoflagger /data/{ms_file}"
-        subprocess.run(aoflag_cmd, shell=True, executable=bash_exec)
-        print()
+        write_to_aoflags(ms_path)
 
     flags = xds_from_ms(ms_path)[0].FLAG.data
     print(f"Flag Rate      : {100*flags.mean().compute(): .1f} %")
@@ -106,17 +138,27 @@ def main():
         "--n_sigma",
         default=3.0,
         type=float,
-        help="Threshold in number of std of noise given by SIGMA column. 0 unflags everything.",
+        help="Threshold in number of std of noise given by SIGMA column. 0 unflags everything. Default is 3.0.",
     )
     parser.add_argument(
-        "-d", "--data_col", default="DATA", help="Data column to run AOFlagger on."
+        "-d",
+        "--data_col",
+        default="DATA",
+        help="Data column to run AOFlagger on. Default is 'DATA'.",
     )
     parser.add_argument(
         "-ao",
         "--aoflagger",
-        default=None,
+        default=False,
         action=argparse.BooleanOptionalAction,
-        help="Whether to use AOFlagger to flag the data.",
+        help="Whether to use AOFlagger to flag the data. Default is False.",
+    )
+    parser.add_argument(
+        "-rao",
+        "--rerun_aoflagger",
+        default=False,
+        action=argparse.BooleanOptionalAction,
+        help="Whether to rerun AOFlagger to flag the data. Default is False.",
     )
     parser.add_argument(
         "-st",
@@ -136,10 +178,12 @@ def main():
     strategy_paths = args.strategy_paths
     data_col = args.data_col
 
-    if args.aoflagger is not None:
+    if args.aoflagger:
         if strategy_paths is not None:
             strategy_paths = strategy_paths.split(",")
-        flags = run_aoflagger(ms_path, data_col, strategy_paths, args.sif_path)
+        flags = run_aoflagger(
+            ms_path, data_col, strategy_paths, args.sif_path, args.rerun_aoflagger
+        )
     else:
         flags = write_perfect_flags(ms_path, n_sigma)
